Peer Magic Update as Security Countermeasure
From May 4th 2022 to approx May 20th 2022 the network was attacked by a group calling themselves Sumcoin Community.  Stating a fork was to occur, it became necessary to protect the real network from any further damage.  Peer magic value change was decided on and implemented here in "v0.17.3" in order to protect the chain before conversion to POS.